==> 06_Load_database.py <==

# coding: utf-8

# # Part 6. Load database
# App to analyze web-site search logs (internal search)<br>
# **This script:** The log table, in the database<br>
# Authors: dan.wendling@nih.gov, <br>
# Last modified: 2018-09-09
# 
# For now let's load search_log and semantic_network. I decided not to load other tables for now; let's see how the work goes. The next candidate would be 01_Text_wrangling_files/GoldStandard_master.xlsx
# 
# Preference: Postgres. If MySQL, the 03_Fuzzy_match file has code for SQLAlchemy, MySQLConnector, etc.
# 
# 
# # Contents
# 1. search_log table
# 2. manual_assignments table
# 3. semantic_network table

# # 1. search_log table

# In[2]:


'''
sql code; this has worked with past code.

DROP TABLE IF EXISTS `search_log`;
CREATE TABLE `search_log` (
    `search_log_id` INT PRIMARY KEY NOT NULL AUTO_INCREMENT,
    `Timestamp` datetime DEFAULT NULL,
    `Query` varchar(800) DEFAULT NULL,
    `Address` varchar(900) DEFAULT NULL,
    `SessionID` varchar(15) NOT NULL,
    `preferredTerm` text,
    `SemanticTypeName` text,
    `SemanticTypeCode` int(11) DEFAULT NULL,
    `SemanticGroup` text,
    `SemanticGroupCode` int(11) DEFAULT NULL,
    `Month` text
) ENGINE=InnoDB DEFAULT CHARSET=utf8;

# For a quick start with data, please try 06_Load_database/logAfterGoldStandard.xlsx, 
which was copied over from 01_Text_wrangling_files.
'''


# # 2. manual_assignments table
# 
# See 03_Fuzzy_match for how this is constructed/used.
# 
# <img src="03_Fuzzy_match_files/fuzzyMatch-Browser.png" />

# In[ ]:


'''    
DROP TABLE IF EXISTS manual_assignments;
CREATE TABLE `manual_assignments` (
  `assignment_id` INT PRIMARY KEY NOT NULL AUTO_INCREMENT,
  `adjustedQueryCase` varchar(200) NULL,
  `NewSemanticTypeName` varchar(100) NULL,
  `preferredTerm` varchar(200) NULL,
  `FuzzyToken` varchar(50) NULL,
  `SemanticTypeName` varchar(100) NULL,
  `SemanticGroup` varchar(50) NULL,
  `timesSearched` int(11) NULL,
  `FuzzyScore` int(11) NULL
) ENGINE=InnoDB DEFAULT CHARSET=utf8;
'''


# # 3. semantic_network table
# I didn't see what I needed online, so created it here. Should be sufficient for joining with the processed logs for reporting.
# 
# Table has one UMLS Semantic Type per row.
# 
# * SemanticGroupCode: With SemanticGroup, SemanticGroupAbr, identifies ~15 supergroups, see McCray AT, Burgun A, Bodenreider O. (2001). Aggregating UMLS semantic types for reducing conceptual complexity. Stud Health Technol Inform. 84(Pt 1):216-20. PMID: 11604736. http://www.ncbi.nlm.nih.gov/pmc/articles/PMC4300099/ and https://semanticnetwork.nlm.nih.gov/.
# * SemanticGroup: See SemanticGroupCode.
# * SemanticGroupAbr: See SemanticGroupCode.
# * CustomTreeNumber: An attempt to get queries to dump in the correct order so counts could be attached to each semantic type, with proper indentation.
# * SemanticTypeName: See https://www.nlm.nih.gov/research/umls/META3_current_semantic_types.html
# * BranchPosition: Use to create indents in browser-based reporting to make it look like https://www.nlm.nih.gov/research/umls/META3_current_semantic_types.html (but one column).
# * Definition: Sem type Definition from UMLS documentation.
# * Examples: Sem type examples from UMLS doc, with a few added.
# * RelationName: Semantic "triples" from UMLS doc; not currently used.
# * SemTypeTreeNo: Sem type tree number from UMLS doc; not currently used.
# * UsageNote: From UMLS doc.
# * Abbreviation: Sem type abbrev from UMLS doc; not currently used.
# * UniqueID: Another attempt to sort the table in hierarchical order.
# * NonHumanFlag: From UMLS doc; not currently used.
# * RecordType: From UMLS doc; not currently used. The UMLS Semantic Network includes other content that could be added such as if we wanted to do {item} {howrelated} {item}. 
# 
# This information can go out of date over time. It is current as of Summer 2018.

# In[ ]:


'''
# sql code
CREATE TABLE `semantic_network` (
  `SemanticGroupCode` bigint(20) DEFAULT NULL,
  `SemanticGroup` text,
  `SemanticGroupAbr` text,
  `CustomTreeNumber` bigint(20) DEFAULT NULL,
  `SemanticTypeName` text,
  `BranchPosition` bigint(20) DEFAULT NULL,
  `Definition` text,
  `Examples` text,
  `RelationName` text,
  `SemTypeTreeNo` text,
  `UsageNote` text,
  `Abbreviation` text,
  `UniqueID` bigint(20) DEFAULT NULL,
  `NonHumanFlag` text,
  `RecordType` text
) ENGINE=InnoDB DEFAULT CHARSET=utf8;

--
-- Dumping data for table `semantic_network`
--

INSERT INTO `semantic_network` (`SemanticGroupCode`, `SemanticGroup`, `SemanticGroupAbr`, `CustomTreeNumber`, `SemanticTypeName`, `BranchPosition`, `Definition`, `Examples`, `RelationName`, `SemTypeTreeNo`, `UsageNote`, `Abbreviation`, `UniqueID`, `NonHumanFlag`, `RecordType`) VALUES
(1, 'Activities and Behaviors', 'ACTI', 2, 'Event', 1, 'A broad type for grouping activities, processes and states.', 'Anniversaries; Exposure to Mumps virus (event); Device Unattended', '{inverse_isa} Activity; {inverse_isa} Phenomenon or Process', 'B', 'Few concepts will be assigned to this broad type.', 'evnt', 1051, NULL, 'STY'),
(1, 'Activities and Behaviors', 'ACTI', 21, 'Activity', 2, 'An operation or series of operations that an organism or machine carries out or participates in.', 'Expeditions; Information Distribution; Social Planning', '{isa} Event; {inverse_isa} Behavior; {inverse_isa} Daily or Recreational Activity; {inverse_isa} Occupational Activity; {inverse_isa} Machine Activity', 'B1', 'Few concepts will be assigned to this broad type. Wherever possible, one of the more specific types from this hierarchy will be chosen. For concepts assigned to this type, the focus of interest is on the activity. When the focus of interest is the individual or group that is carrying out the activity, then a type from the \'Behavior\' hierarchy will be chosen. In general, concepts will not receive a type from both the \'Activity\' and the \'Behavior\' hierarchies.', 'acty', 1052, NULL, 'STY'),
(1, 'Activities and Behaviors', 'ACTI', 211, 'Behavior', 3, 'Any of the psycho-social activities of humans or animals that can be observed directly by others or can be made systematically observable by the use of special strategies.', 'Homing Behavior; Sexuality; Habitat Selection', '{isa} Activity; {inverse_isa} Social Behavior; {inverse_isa} Individual Behavior', 'B1.1', 'Few concepts will be assigned to this broad type. For concepts assigned to the \'Behavior\' hierarchy, the focus of interest is on the individual or group that is carrying out the activity. When the activity is of paramount interest, then a type from the \'Activity\' hierarchy will be chosen. In general, concepts will not receive a type from both the \'Behavior\' and the \'Activity\' hierarchies.', 'bhvr', 1053, 'Y', 'STY'),
(1, 'Activities and Behaviors', 'ACTI', 212, 'Daily or Recreational Activity', 3, 'An activity carried out for recreation or exercise, or as part of daily life.', 'Badminton; Dancing; Swimming', '{isa} Activity', 'B1.2', NULL, 'dora', 1056, NULL, 'STY'),
(1, 'Activities and Behaviors', 'ACTI', 213, 'Occupational Activity', 3, 'An activity carried out as part of an occupation or job.', 'Collective Bargaining; Commerce; Containment of Biohazards', '{isa} Activity; {inverse_isa} Health Care Activity; {inverse_isa} Research Activity; {inverse_isa} Governmental or Regulatory Activity; {inverse_isa} Educational Activity', 'B1.3', NULL, 'ocac', 1057, NULL, 'STY'),
(1, 'Activities and Behaviors', 'ACTI', 214, 'Machine Activity', 3, 'An activity carried out primarily or exclusively by machines.', 'Computer Simulation; Equipment Failure; Natural Language Processing', '{isa} Activity', 'B1.4', NULL, 'mcha', 1066, NULL, 'STY'),
(1, 'Activities and Behaviors', 'ACTI', 2111, 'Social Behavior', 4, 'Behavior that is a direct result or function of the interaction of humans or animals with their fellows. This includes behavior that may be considered anti-social.', 'Acculturation; Communication; Interpersonal Relations', '{isa} Behavior', 'B1.1.1', '\'Social Behavior\' requires the direct participation of others and is, thus, distinguished from \'Individual Behavior\' which is carried out by an individual, though others may be present.', 'socb', 1054, NULL, 'STY'),
(1, 'Activities and Behaviors', 'ACTI', 2112, 'Individual Behavior', 4, 'Behavior exhibited by a human or an animal that is not a direct result of interaction with other members of the species, but which may have an effect on others.', 'Assertiveness; Grooming; Risk-Taking', '{isa} Behavior', 'B1.1.2', '\'Individual Behavior\' is carried out by an individual, though others may be present, and is, thus, distinguished from \'Social Behavior\' which requires the direct participation of others.', 'inbe', 1055, NULL, 'STY'),
(1, 'Activities and Behaviors', 'ACTI', 2133, 'Governmental or Regulatory Activity', 4, 'An activity carried out by officially constituted governments, or an activity related to the creation or enforcement of the rules or regulations governing some field of endeavor.', 'Certification; Credentialing; Public Policy', '{isa} Occupational Activity', 'B1.3.3', NULL, 'gora', 1064, NULL, 'STY'),
(2, 'Anatomy ', 'ANAT', 112, 'Anatomical Structure', 3, 'A normal or pathological part of the anatomy or structural organization of an organism.', 'Cadaver; Pharyngostome; Anatomic structures', '{isa} Physical Object; {inverse_isa} Embryonic Structure; {inverse_isa} Fully Formed Anatomical Structure; {inverse_isa} Anatomical Abnormality', 'A1.2', 'Few concepts will be assigned to this broad type.', 'anst', 1017, 'Y', 'STY'),
(2, 'Anatomy ', 'ANAT', 1121, 'Embryonic Structure', 4, 'An anatomical structure that exists only before the organism is fully formed; in mammals, for example, a structure that exists only prior to the birth of the organism. This structure may be normal or abnormal.', 'Blastoderm; Fetus; Neural Crest', '{isa} Anatomical Structure', 'A1.2.1', NULL, 'emst', 1018, NULL, 'STY'),
(2, 'Anatomy ', 'ANAT', 1123, 'Fully Formed Anatomical Structure', 4, 'An anatomical structure in a fully formed organism; in mammals, for example, a structure in the body after the birth of the organism.', 'Entire body as a whole; Female human body; Set of parts of human body', '{isa} Anatomical Structure; {inverse_isa} Body Part, Organ, or Organ Component; {inverse_isa} Tissue; {inverse_isa} Cell; {inverse_isa} Cell Component; {inverse_isa} Gene or Genome', 'A1.2.3', 'Few concepts will be assigned to this broad type.', 'ffas', 1021, NULL, 'STY'),
(2, 'Anatomy ', 'ANAT', 1142, 'Body Substance', 4, 'Extracellular material, or mixtures of cells and extracellular material, produced, excreted, or accreted by the body. Included here are substances such as saliva, dental enamel, sweat, and gastric acid.', 'Amniotic Fluid; saliva; Smegma', '{isa} Substance', 'A1.4.2', NULL, 'bdsu', 1031, 'Y', 'STY'),
(2, 'Anatomy ', 'ANAT', 11231, 'Body Part, Organ, or Organ Component', 5, 'A collection of cells and tissues which are localized to a specific area or combine and carry out one or more specialized functions of an organism. This ranges from gross structures to small components of complex organs. These structures are relatively localized in comparison to tissues.', 'Aorta; Brain Stem; Structure of neck of femur', '{isa} Fully Formed Anatomical Structure', 'A1.2.3.1', 'When assigning this type, consider whether \'Body Location or Region\' might be the correct choice.', 'bpoc', 1023, NULL, 'STY'),
(2, 'Anatomy ', 'ANAT', 11232, 'Tissue', 5, 'An aggregation of similarly specialized cells and the associated intercellular substance. Tissues are relatively non-localized in comparison to body parts, organs or organ components.', 'Cartilage; Endothelium; Epidermis', '{isa} Fully Formed Anatomical Structure', 'A1.2.3.2', NULL, 'tisu', 1024, NULL, 'STY'),
(2, 'Anatomy ', 'ANAT', 11233, 'Cell', 5, 'The fundamental structural and functional unit of living organisms.', 'B-Lymphocytes; Dendritic Cells; Fibroblasts', '{isa} Fully Formed Anatomical Structure', 'A1.2.3.3', NULL, 'cell', 1025, NULL, 'STY'),
(2, 'Anatomy ', 'ANAT', 11234, 'Cell Component', 5, 'A part of a cell or the intercellular matrix, generally visible by light microscopy.', 'Axon; Golgi Apparatus; Organelles', '{isa} Fully Formed Anatomical Structure', 'A1.2.3.4', NULL, 'celc', 1026, NULL, 'STY'),
(2, 'Anatomy ', 'ANAT', 12141, 'Body System', 5, 'A complex of anatomical structures that performs a common function.', 'Endocrine system; Renin-angiotensin system; Reticuloendothelial System', '{isa} Functional Concept', 'A2.1.4.1', NULL, 'bdsy', 1022, NULL, 'STY'),
(2, 'Anatomy ', 'ANAT', 12151, 'Body Space or Junction', 5, 'An area enclosed or surrounded by body parts or organs or the place where two anatomical structures meet or connect.', 'Knee joint; Greater sac of peritoneum; Synapses', '{isa} Spatial Concept', 'A2.1.5.1', NULL, 'bsoj', 1030, 'Y', 'STY'),
(2, 'Anatomy ', 'ANAT', 12152, 'Body Location or Region', 5, 'An area, subdivision, or region of the body demarcated for the purpose of topographical description.', 'Forehead; Sublingual Region; Base of skull structure', '{isa} Spatial Concept', 'A2.1.5.2', 'When assigning this type, consider whether \'Body Part, Organ, or Organ Component\' might be the correct choice.', 'blor', 1029, 'Y', 'STY'),
(3, 'Chemicals and Drugs', 'CHEM', 1133, 'Clinical Drug', 4, 'A pharmaceutical preparation as produced by the manufacturer. The name usually includes the substance, its strength, and the form, but may include the substance and only one of the other two items.', 'Ranitidine 300 MG Oral Tablet [Zantac]; Aspirin 300 MG Delayed Release Oral Tablet; sleeping pill', '{isa} Manufactured Object', 'A1.3.3', 'Do not double type with Pharmacologic Substance, Antibiotic, or other chemical semantic types.', 'clnd', 1200, NULL, 'STY'),
(3, 'Chemicals and Drugs', 'CHEM', 1141, 'Chemical', 4, 'Compounds or substances of definite molecular composition. Chemicals are viewed from two distinct perspectives in the network, functionally and structurally. Almost every chemical concept is assigned at least two types, generally one from the structure hierarchy and at least one from the function hierarchy.', 'Acids; Chemicals; Ionic Liquids', '{isa} Substance; {inverse_isa} Chemical Viewed Structurally; {inverse_isa} Chemical Viewed Functionally', 'A1.4.1', 'Few concepts will be assigned to this broad type.', 'chem', 1103, NULL, 'STY'),
(3, 'Chemicals and Drugs', 'CHEM', 11411, 'Chemical Viewed Functionally', 5, 'A chemical viewed from the perspective of its functional characteristics or pharmacological activities.', 'Aerosol Propellants; Detergents; Stabilizing Agents', '{isa} Chemical; {inverse_isa} Pharmacologic Substance; {inverse_isa} Biomedical or Dental Material; {inverse_isa} Biologically Active Substance; {inverse_isa} Indicator, Reagent, or Diagnostic Aid; {inverse_isa} Hazardous or Poisonous Substance', 'A1.4.1.1', 'A specific chemical will not be assigned here. Groupings of chemicals viewed functionally, such as \"Aerosol Propellants\" may appropriately be assigned here. A name that is inherently functional, such as \"Food Additives\", will not also be assigned a type from the \'Chemical Viewed Structurally\' hierarchy.', 'chvf', 1120, NULL, 'STY'),
(3, 'Chemicals and Drugs', 'CHEM', 11412, 'Chemical Viewed Structurally', 5, 'A chemical or chemicals viewed from the perspective of their structural characteristics. Included here are concepts which can mean either a salt, an ion, or a compound (e.g., \"Bromates\" and \"Bromides\").', 'Ammonium Compounds; Cations; Sulfur Compounds', '{isa} Chemical; {inverse_isa} Organic Chemical; {inverse_isa} Element, Ion, or Isotope; {inverse_isa} Inorganic Chemical', 'A1.4.1.2', 'Concepts are assigned to this type if they can be both organic and inorganic, e.g. sulfur compounds. Do not use this type if the concept has an important functional aspect, e.g., \"Mylanta Double Strength Liquid\" contains Al(OH)3, Mg(OH)2, and simethicone, but would be assigned only to \'Pharmacologic Substance\'.', 'chvs', 1104, NULL, 'STY'),
(3, 'Chemicals and Drugs', 'CHEM', 114111, 'Pharmacologic Substance', 6, 'A substance used in the treatment or prevention of pathologic disorders. This includes substances that occur naturally in the body and are administered therapeutically.', 'Antiemetics; Cardiovascular Agents; Alka-Seltzer', '{isa} Chemical Viewed Functionally; {inverse_isa} Antibiotic', 'A1.4.1.1.1', 'If a substance is both endogenous and typically used as a drug, then this type and the type \'Biologically Active Substance\' or one of its children are assigned. Body substances that are used therapeutically such as whole blood preparation, NOS would only receive the type \'Body Substance\'. Substances used in the diagnosis or analysis of normal and abnormal body functions should be given the type \'Indicator, Reagent, or Diagnostic Aid\'.', 'phsu', 1121, NULL, 'STY'),
(3, 'Chemicals and Drugs', 'CHEM', 114112, 'Biomedical or Dental Material', 6, 'A substance used in biomedicine or dentistry predominantly for its physical, as opposed to chemical, properties. Included here are biocompatible materials, tissue adhesives, bone cements, resins, toothpastes, etc.', 'Acrylic Resins; Bone Cements; Dentifrices', '{isa} Chemical Viewed Functionally', 'A1.4.1.1.2', NULL, 'bodm', 1122, NULL, 'STY'),
(3, 'Chemicals and Drugs', 'CHEM', 114113, 'Biologically Active Substance', 6, 'A generally endogenous substance produced or required by an organism, of primary interest because of its role in the biologic functioning of the organism that produces it.', 'Cytokinins; Pheromone', '{isa} Chemical Viewed Functionally; {inverse_isa} Hormone; {inverse_isa} Enzyme; {inverse_isa} Vitamin; {inverse_isa} Immunologic Factor; {inverse_isa} Receptor', 'A1.4.1.1.3', 'If a substance is both endogenous and typically used as a drug, then this type and the type \'Pharmacologic Substance\' are assigned.', 'bacs', 1123, NULL, 'STY'),
(3, 'Chemicals and Drugs', 'CHEM', 114114, 'Indicator, Reagent, or Diagnostic Aid', 6, 'A substance primarily of interest for its use in laboratory or diagnostic tests and procedures to detect, measure, examine, or analyze other chemicals, processes, or conditions.', 'Fluorescent Dyes; Indicators and Reagents; India ink stain', '{isa} Chemical Viewed Functionally', 'A1.4.1.1.4', 'Radioactive imaging agents should be assigned to this type and not to the type \'Pharmacologic Substance\' unless they are also being used therapeutically.', 'irda', 1130, NULL, 'STY'),
(3, 'Chemicals and Drugs', 'CHEM', 114115, 'Hazardous or Poisonous Substance', 6, 'A substance of concern because of its potentially hazardous or toxic effects. This would include most drugs of abuse, as well as agents that require special handling because of their toxicity.', 'Carcinogens; Fumigant; Mutagens', '{isa} Chemical Viewed Functionally', 'A1.4.1.1.5', 'Most pharmaceutical agents, although potentially harmful, are excluded here and are assigned to the type \'Pharmacologic Substance\'. All pesticides are assigned to this type.', 'hops', 1131, NULL, 'STY'),
(3, 'Chemicals and Drugs', 'CHEM', 114121, 'Organic Chemical', 6, 'The general class of carbon-containing compounds, usually based on carbon chains or rings, and also containing hydrogen (hydrocarbons), with or without nitrogen, oxygen, or other elements in which the bonding between elements is generally covalent.', 'Benzene Derivatives', '{isa} Chemical Viewed Structurally; {inverse_isa} Nucleic Acid, Nucleoside, or Nucleotide; {inverse_isa} Amino Acid, Peptide, or Protein', 'A1.4.1.2.1', 'Salts of organic chemicals (such as Calcium Acetate) would be considered organic chemicals and should not also receive the type \'Inorganic Chemical\'.', 'orch', 1109, NULL, 'STY'),
(3, 'Chemicals and Drugs', 'CHEM', 114122, 'Inorganic Chemical', 6, 'Chemical elements and their compounds, excluding the hydrocarbons and their derivatives (except carbides, carbonates, cyanides, cyanates and carbon disulfide). Generally inorganic compounds contain ionic bonds. Included here are inorganic acids and salts, alloys, alkalies, and minerals.', 'Carbonic Acid; aluminum nitride; ferric citrate', '{isa} Chemical Viewed Structurally', 'A1.4.1.2.2', NULL, 'inch', 1197, NULL, 'STY'),
(3, 'Chemicals and Drugs', 'CHEM', 114123, 'Element, Ion, or Isotope', 6, 'One of the 109 presently known fundamental substances that comprise all matter at and above the atomic level. This includes elemental metals, rare gases, and most abundant naturally occurring radioactive elements, as well as the ionic counterparts of elements (NA+, Cl-), and the less abundant isotopic forms. This does not include organic ions such as iodoacetate to which the type \'Organic Chemical\' is assigned.', 'Carbon; Chromium Isotopes; Radioisotopes', '{isa} Chemical Viewed Structurally', 'A1.4.1.2.3', 'Group terms such as sulfates would be assigned to the type \'Chemical Viewed Structurally\'. Substances such as aluminum chloride would be assigned the type \'Inorganic Chemical\'. Technetium Tc 99m Aggregated Albumin would not receive this type.', 'elii', 1196, NULL, 'STY'),
(3, 'Chemicals and Drugs', 'CHEM', 1141111, 'Antibiotic', 7, 'A pharmacologically active compound produced by growing microorganisms which kill or inhibit growth of other microorganisms.', 'Antibiotics; bactericide; Thienamycins', '{isa} Pharmacologic Substance', 'A1.4.1.1.1.1', NULL, 'antb', 1195, NULL, 'STY'),
(3, 'Chemicals and Drugs', 'CHEM', 1141132, 'Hormone', 7, 'In animals, a chemical usually secreted by an endocrine gland whose products are released into the circulating fluid. Hormones act as chemical messengers and regulate various physiologic processes such as growth, reproduction, metabolism, etc. They usually fall into two broad classes, steroid hormones and peptide hormones.', 'Enteric Hormones; thymic humoral factor; Prohormone', '{isa} Biologically Active Substance', 'A1.4.1.1.3.2', 'Synthetic hormones that are used as drugs should receive this type and \'Pharmacologic Substance\'. Plant hormones are assigned only to the type \'Pharmacologic Substance\'.', 'horm', 1125, NULL, 'STY'),
(3, 'Chemicals and Drugs', 'CHEM', 1141133, 'Enzyme', 7, 'A complex chemical, usually a protein, that is produced by living cells and which catalyzes specific biochemical reactions. There are six main types of enzymes: oxidoreductases, transferases, hydrolases, lyases, isomerases, and ligases.', 'GTP Cyclohydrolase II; enzyme substrate complex; arginine amidase', '{isa} Biologically Active Substance', 'A1.4.1.1.3.3', 'Generally when a concept is assigned to this type, it will also be assigned to the type \'Amino Acid, Peptide, or Protein\'.', 'enzy', 1126, NULL, 'STY'),
(3, 'Chemicals and Drugs', 'CHEM', 1141134, 'Vitamin', 7, 'A substance, usually an organic chemical complex, present in natural products or made synthetically, which is essential in the diet of man or other higher animals. Included here are vitamin precursors, provitamins, and vitamin supplements.', '5,25-Dihydroxy cholecalciferol; alpha-tocopheryl oxalate; Vitamin A [EPC]', '{isa} Biologically Active Substance', 'A1.4.1.1.3.4', 'Essential amino acids are not assigned to this type. They will be assigned to the type \'Amino Acid, Peptide, or Protein\'. This can be used with \'Pharmacologic Substance\' if the compound is being administered therapeutically or if the source has it classified as therapeutic (i.e., N\'ICE Sugarless Vitamin C Drops).', 'vita', 1127, NULL, 'STY'),
(3, 'Chemicals and Drugs', 'CHEM', 1141135, 'Immunologic Factor', 7, 'A biologically active substance whose activities affect or play a role in the functioning of the immune system.', 'Antigens; Immunologic Factors; Blood group antigen P', '{isa} Biologically Active Substance', 'A1.4.1.1.3.5', 'Antigens and antibodies are assigned to this type. Unlike most biologically active substances, some immunologic factors may be exogenous. Vaccines should be given this type and the type \'Pharmacologic Substance\'.', 'imft', 1129, NULL, 'STY'),
(3, 'Chemicals and Drugs', 'CHEM', 1141136, 'Receptor', 7, 'A specific structure or site on the cell surface or within its cytoplasm that recognizes and binds with other specific molecules. These include the proteins on the surface of an immunocompetent cell that binds with antigens, or proteins found on the surface molecules that bind with hormones or neurotransmitters and react with other molecules that respond in a specific way.', 'Binding Sites; Lymphocyte antigen CD4 receptor; integrin alpha11beta1', '{isa} Biologically Active Substance', 'A1.4.1.1.3.6', NULL, 'rcpt', 1192, NULL, 'STY'),
(3, 'Chemicals and Drugs', 'CHEM', 1141215, 'Nucleic Acid, Nucleoside, or Nucleotide', 7, 'A complex compound of high molecular weight occurring in living cells. These are basically of two types, ribonucleic (RNA) and deoxyribonucleic (DNA) acids. Nucleic acids are made of nucleotides (nitrogen-containing base, a 5-carbon sugar, and one or more phosphate group) linked together by a phosphodiester bond between the 5\' and 3\' carbon atoms. Nucleosides are compounds composed of a purine or pyrimidine base (usually adenine, cytosine, guanine, thymine, uracil) linked to either a ribose or a deoxyribose sugar.', 'Cytosine Nucleotides; Guanine; Oligonucleotides', '{isa} Organic Chemical', 'A1.4.1.2.1.5', 'Naturally occurring nucleic acids, nucleosides, or nucleotides will also be assigned a type from the \'Biologically Active Substance\' hierarchy.', 'nnon', 1114, NULL, 'STY'),
(3, 'Chemicals and Drugs', 'CHEM', 1141217, 'Amino Acid, Peptide, or Protein', 7, 'Amino acids and chains of amino acids connected by peptide linkages.', 'Amino Acids, Cyclic; Glycopeptides; Keratin', '{isa} Organic Chemical', 'A1.4.1.2.1.7', 'When the concept is both an enzyme and a protein, this type and the type \'Enzyme\' will be assigned.', 'aapp', 1116, NULL, 'STY'),
(4, 'Concepts and Ideas', 'CONC', 12, 'Conceptual Entity', 2, 'A broad type for grouping abstract entities or concepts.', 'Geographic Factors; Fractals; Secularism', '{isa} Entity; {inverse_isa} Organism Attribute; {inverse_isa} Finding; {inverse_isa} Idea or Concept; {inverse_isa} Occupation or Discipline; {inverse_isa} Organization; {inverse_isa} Group; {inverse_isa} Group Attribute; {inverse_isa} Intellectual Product; {inverse_isa} Language', 'A2', 'Few concepts will be assigned to this broad type.', 'cnce', 1077, NULL, 'STY'),
(4, 'Concepts and Ideas', 'CONC', 121, 'Idea or Concept', 3, 'An abstract concept, such as a social, religious or philosophical concept.', 'Capitalism; Civil Rights; Ethics', '{isa} Conceptual Entity; {inverse_isa} Temporal Concept; {inverse_isa} Qualitative Concept; {inverse_isa} Quantitative Concept; {inverse_isa} Spatial Concept; {inverse_isa} Functional Concept', 'A2.1', NULL, 'idcn', 1078, NULL, 'STY'),
(4, 'Concepts and Ideas', 'CONC', 124, 'Intellectual Product', 3, 'A conceptual entity resulting from human endeavor. Concepts assigned to this type generally refer to information created by humans for some purpose.', 'Decision Support Techniques; Information Systems; Literature', '{isa} Conceptual Entity; {inverse_isa} Regulation or Law; {inverse_isa} Classification', 'A2.4', 'Concepts referring to theorems, models, and systems are assigned here. In some cases, a concept may be assigned to both \'Intellectual Product\' and \'Research Activity\'. For example, the concept \"Comparative Study\" might be viewed as both an activity and the result, or product, of that activity.', 'inpr', 1170, NULL, 'STY'),
(4, 'Concepts and Ideas', 'CONC', 125, 'Language', 3, 'The system of communication used by a particular nation or people.', 'Armenian language; braille; Bilingualism', '{isa} Conceptual Entity', 'A2.5', NULL, 'lang', 1171, NULL, 'STY'),
(4, 'Concepts and Ideas', 'CONC', 128, 'Group Attribute', 3, 'A conceptual entity which refers to the frequency or distribution of certain characteristics or phenomena in certain groups.', 'Family Size; Group Structure; Life Expectancy', '{isa} Conceptual Entity', 'A2.8', NULL, 'grpa', 1102, NULL, 'STY'),
(4, 'Concepts and Ideas', 'CONC', 1211, 'Temporal Concept', 4, 'A concept which pertains to time or duration.', 'Birth Intervals; Half-Life; Postoperative Period', '{isa} Idea or Concept', 'A2.1.1', 'If the concept refers to a phase, stage, cycle, interval, period, or rhythm, it is assigned to this type.', 'tmco', 1079, NULL, 'STY'),
(4, 'Concepts and Ideas', 'CONC', 1212, 'Qualitative Concept', 4, 'A concept which is an assessment of some quality, rather than a direct measurement.', 'Clinical Competence; Consumer Satisfaction; Health Status', '{isa} Idea or Concept', 'A2.1.2', NULL, 'qlco', 1080, NULL, 'STY'),
(4, 'Concepts and Ideas', 'CONC', 1213, 'Quantitative Concept', 4, 'A concept which involves the dimensions, quantity or capacity of something using some unit of measure, or which involves the quantitative comparison of entities.', 'Age Distribution; Metric System; Selection Bias', '{isa} Idea or Concept', 'A2.1.3', 'If the concept refers to rate or distribution, the type \'Temporal Concept\' is not also assigned.', 'qnco', 1081, NULL, 'STY'),
(4, 'Concepts and Ideas', 'CONC', 1214, 'Functional Concept', 4, 'A concept which is of interest because it pertains to the carrying out of a process or activity.', 'Interviewer Effect; Problem Formulation; Endogenous', '{isa} Idea or Concept; {inverse_isa} Body System', 'A2.1.4', NULL, 'ftcn', 1169, NULL, 'STY'),
(4, 'Concepts and Ideas', 'CONC', 1215, 'Spatial Concept', 4, 'A location, region, or space, generally having definite boundaries.', 'Mandibular Rest Position; Lateral; Extrinsic', '{isa} Idea or Concept; {inverse_isa} Body Location or Region; {inverse_isa} Body Space or Junction; {inverse_isa} Geographic Area; {inverse_isa} Molecular Sequence', 'A2.1.5', NULL, 'spco', 1082, NULL, 'STY'),
(4, 'Concepts and Ideas', 'CONC', 1241, 'Classification', 4, 'A term or system of terms denoting an arrangement by class or category.', 'Anatomy (MeSH Category); Tumor Stage Classification; axis i', '{isa} Intellectual Product', 'A2.4.1', NULL, 'clas', 1185, NULL, 'STY'),
(4, 'Concepts and Ideas', 'CONC', 1242, 'Regulation or Law', 4, 'An intellectual product resulting from legislative or regulatory activity.', 'Building Codes; Criminal Law; Health Planning Guidelines', '{isa} Intellectual Product', 'A2.4.2', NULL, 'rnlw', 1089, NULL, 'STY'),
(5, 'Devices', 'DEVI', 1131, 'Medical Device', 4, 'A manufactured object used primarily in the diagnosis, treatment, or prevention of physiologic or anatomic disorders.', 'Bone Screws; Headgear, Orthodontic; Compression Stockings', '{isa} Manufactured Object; {inverse_isa} Drug Delivery Device', 'A1.3.1', 'A medical device may be used for research purposes, but since its primary use is for routine medical care, it is distinguished from a \'Research Device\' which is used primarily for research purposes.', 'medd', 1074, NULL, 'STY'),
(5, 'Devices', 'DEVI', 1132, 'Research Device', 4, 'A manufactured object used primarily in carrying out scientific research or experimentation.', 'Electrodes, Enzyme; DNA Microarray Chip; Particle Count and Size Analyzer', '{isa} Manufactured Object', 'A1.3.2', 'A research device is distinguished from a \'Medical Device\', which though it may be used for research purposes is used primarily for routine medical care.', 'resd', 1075, NULL, 'STY'),
(5, 'Devices', 'DEVI', 11311, 'Drug Delivery Device', 5, 'A medical device that contains a clinical drug or drugs.', 'Nordette 21 Day Pack; {7 (Terazosin 1 MG Oral Tablet) / 7 (Terazosin 2 MG Oral Tablet) } Pack; {10 (cefdinir 300 MG Oral Capsule [Omnicef]) } Pack [Omni-Pac]', '{isa} Medical Device', 'A1.3.1.1', NULL, 'drdd', 1203, NULL, 'STY'),
(6, 'Disorders', 'DISO', 122, 'Finding', 3, 'That which is discovered by direct observation or measurement of an organism attribute or condition, including the clinical history of the patient. The history of the presence of a disease is a \'Finding\' and is distinguished from the disease itself.', 'Birth History; Downward displacement of diaphragm; Decreased glucose level', '{isa} Conceptual Entity; {inverse_isa} Laboratory or Test Result; {inverse_isa} Sign or Symptom', 'A2.2', 'Only in rare circumstances will findings be double-typed with either \'Pathologic Function\' or \'Anatomical Abnormality\'. Most findings will be assigned the types \'Laboratory or Test Result\' or \'Sign or Symptom\'. Only those findings that relate to patient history or to the determination of a state will be assigned the type \'Finding\'.', 'fndg', 1033, NULL, 'STY'),
(6, 'Disorders', 'DISO', 223, 'Injury or Poisoning', 3, 'A traumatic wound, injury, or poisoning caused by an external agent or force.', 'Accidental Falls; Carbon Monoxide Poisoning; Snake Bites', '{isa} Phenomenon or Process', 'B2.3', 'An `Injury or Poisoning\' is distinguished from a \'Disease or Syndrome\' that may be a result of prolonged exposure to toxic materials.', 'inpo', 1037, NULL, 'STY'),
(6, 'Disorders', 'DISO', 1122, 'Anatomical Abnormality', 4, 'An abnormal structure, or one that is abnormal in size or location.', 'Bronchial Fistula; Foot Deformities; Hyperostosis of skull', '{isa} Anatomical Structure; {inverse_isa} Congenital Abnormality; {inverse_isa} Acquired Abnormality', 'A1.2.2', 'Use this type if the abnormality in question can be either an acquired or congenital abnormality. Neoplasms are not included here. These are given the type \'Neoplastic Process\'. If an anatomical abnormality has a pathologic manifestation, then it will additionally be given the type \'Disease or Syndrome\', e.g., \"Diabetic Cataract\" will be double-typed for this reason.', 'anab', 1190, NULL, 'STY'),
(6, 'Disorders', 'DISO', 1222, 'Sign or Symptom', 4, 'An observable manifestation of a disease or condition based on clinical judgment, or a manifestation of a disease or condition which is experienced by the patient and reported as a subjective observation.', 'Dyspnea; Nausea; Pain', '{isa} Finding', 'A2.2.2', NULL, 'sosy', 1184, NULL, 'STY'),
(6, 'Disorders', 'DISO', 11221, 'Congenital Abnormality', 5, 'An abnormal structure, or one that is abnormal in size or location, present at birth or evolving over time as a result of a defect in embryogenesis.', 'Albinism; Cleft palate with cleft lip; Polydactyly of toes', '{isa} Anatomical Abnormality', 'A1.2.2.1', 'If the congenital abnormality involves multiple defects then the type \'Disease or Syndrome\' will also be assigned.', 'cgab', 1019, NULL, 'STY'),
(6, 'Disorders', 'DISO', 11222, 'Acquired Abnormality', 5, 'An abnormal structure, or one that is abnormal in size or location, found in or deriving from a previously normal structure. Acquired abnormalities are distinguished from diseases even though they may result in pathological functioning (e.g., \"hernias incarcerate\").', 'Hemorrhoids; Hernia, Femoral; Cauliflower ear', '{isa} Anatomical Abnormality', 'A1.2.2.2', NULL, 'acab', 1020, NULL, 'STY'),
(6, 'Disorders', 'DISO', 22212, 'Pathologic Function', 5, 'A disordered process, activity, or state of the organism as a whole, of a body system or systems, or of multiple organs or tissues. Included here are normal responses to a negative stimulus as well as patholologic conditions or states that are less specific than a disease. Pathologic functions frequently have systemic effects.', 'Inflammation; Shock; Thrombosis', '{isa} Biologic Function; {inverse_isa} Disease or Syndrome; {inverse_isa} Cell or Molecular Dysfunction; {inverse_isa} Experimental Model of Disease', 'B2.2.1.2', 'If the process is specific, for example to a site or substance, then \'Disease or Syndrome\' will be assigned and not \'Pathologic Function\'. For example, \"cerebral anoxia\", \"brain edema\", and \"milk hypersensitivity\" will all be assigned to \'Disease or Syndrome\' only.', 'patf', 1046, NULL, 'STY'),
(6, 'Disorders', 'DISO', 222121, 'Disease or Syndrome', 6, 'A condition which alters or interferes with a normal process, state, or activity of an organism. It is usually characterized by the abnormal functioning of one or more of the host\'s systems, parts, or organs. Included here is a complex of symptoms descriptive of a disorder.', 'Diabetes Mellitus; Drug Allergy; Malabsorption Syndrome', '{isa} Pathologic Function; {inverse_isa} Mental or Behavioral Dysfunction; {inverse_isa} Neoplastic Process', 'B2.2.1.2.1', 'Any specific disease or syndrome that is modified by such modifiers as \"acute\", \"prolonged\", etc. will also be assigned to this type. If an anatomic abnormality has a pathologic manifestation, then it will be given this type as well as a type from the \'Anatomical Abnormality\' hierarchy, e.g., \"Diabetic Cataract\" will be double-typed for this reason.', 'dsyn', 1047, NULL, 'STY'),
(6, 'Disorders', 'DISO', 222122, 'Cell or Molecular Dysfunction', 6, 'A pathologic function inherent to cells, parts of cells, or molecules.', 'DNA Damage; Wallerian Degeneration; Atypical squamous metaplasia', '{isa} Pathologic Function', 'B2.2.1.2.2', 'This is not intended to be a repository for diseases whose molecular basis has been established.', 'comd', 1049, NULL, 'STY'),
(6, 'Disorders', 'DISO', 222123, 'Experimental Model of Disease', 6, 'A representation in a non-human organism of a human disease for the purpose of research into its mechanism or treatment.', 'Alloxan Diabetes; Liver Cirrhosis, Experimental; Transient Gene Knock-Out Model', '{isa} Pathologic Function', 'B2.2.1.2.3', NULL, 'emod', 1050, NULL, 'STY'),
(6, 'Disorders', 'DISO', 2221211, 'Mental or Behavioral Dysfunction', 7, 'A clinically significant dysfunction whose major manifestation is behavioral or psychological. These dysfunctions may have identified or presumed biological etiologies or manifestations.', 'Agoraphobia; Cyclothymic Disorder; Frigidity', '{isa} Disease or Syndrome', 'B2.2.1.2.1.1', NULL, 'mobd', 1048, NULL, 'STY'),
(6, 'Disorders', 'DISO', 2221212, 'Neoplastic Process', 7, 'A new and abnormal growth of tissue in which the growth is uncontrolled and progressive. The growths may be malignant or benign.', 'Abdominal Neoplasms; Bowen\'s Disease; Polyp in nasopharynx', '{isa} Disease or Syndrome', 'B2.2.1.2.1.2', 'All neoplasms are assigned to this type. Do not also assign a type from the \'Anatomical Abnormality\' hierarchy.', 'neop', 1191, NULL, 'STY'),
(7, 'Genes and Molecular Sequences', 'GENE', 11235, 'Gene or Genome', 5, 'A specific sequence, or in the case of the genome the complete sequence, of nucleotides along a molecule of DNA or RNA (in the case of some viruses) which represent the functional units of heredity.', 'Alleles; Genome, Human; rRNA Operon', '{isa} Fully Formed Anatomical Structure', 'A1.2.3.5', NULL, 'gngm', 1028, NULL, 'STY'),
(7, 'Genes and Molecular Sequences', 'GENE', 12153, 'Molecular Sequence', 5, 'A broad type for grouping the collected sequences of amino acids, carbohydrates, and nucleotide sequences. Descriptions of these sequences are generally reported in the published literature and/or are deposited in and maintained by databanks such as GenBank, European Molecular Biology Laboratory (EMBL), National Biomedical Research Foundation (NBRF), or other sequence repositories.', 'Genetic Code; Homologous Sequences; Molecular Sequence', '{isa} Spatial Concept; {inverse_isa} Nucleotide Sequence; {inverse_isa} Amino Acid Sequence; {inverse_isa} Carbohydrate Sequence', 'A2.1.5.3', NULL, 'mosq', 1085, NULL, 'STY'),
(7, 'Genes and Molecular Sequences', 'GENE', 121531, 'Nucleotide Sequence', 6, 'The sequence of purines and pyrimidines in nucleic acids and polynucleotides. Included here are nucleotide-rich regions, conserved sequence, and DNA transforming region.', 'Base Sequence; Direct Repeat; RNA Sequence', '{isa} Molecular Sequence', 'A2.1.5.3.1', NULL, 'nusq', 1086, NULL, 'STY'),
(7, 'Genes and Molecular Sequences', 'GENE', 121532, 'Amino Acid Sequence', 6, 'The sequence of amino acids as arrayed in chains, sheets, etc., within the protein molecule. It is of fundamental importance in determining protein structure.', 'Signal Peptides; Homologous Sequences, Amino Acid; Abnormal amino acid sequence', '{isa} Molecular Sequence', 'A2.1.5.3.2', NULL, 'amas', 1087, NULL, 'STY'),
(7, 'Genes and Molecular Sequences', 'GENE', 121533, 'Carbohydrate Sequence', 6, 'The sequence of carbohydrates within polysaccharides, glycoproteins, and glycolipids.', 'Carbohydrate Sequence; Abnormal carbohydrate sequence', '{isa} Molecular Sequence', 'A2.1.5.3.3', NULL, 'crbs', 1088, NULL, 'STY'),
(8, 'Geographic Areas', 'GEOG', 12154, 'Geographic Area', 5, 'A geographic location, generally having definite boundaries.', 'Baltimore; Canada; Far East', '{isa} Spatial Concept', 'A2.1.5.4', NULL, 'geoa', 1083, NULL, 'STY'),
(9, 'Living Beings', 'LIVE', 111, 'Organism', 3, 'Generally, a living individual, including all plants and animals.', 'Organism; Infectious agent; Heterotroph', '{isa} Physical Object; {inverse_isa} Virus; {inverse_isa} Bacterium; {inverse_isa} Archaeon; {inverse_isa} Eukaryote', 'A1.1', NULL, 'orgm', 1001, NULL, 'STY'),
(9, 'Living Beings', 'LIVE', 129, 'Group', 3, 'A conceptual entity referring to the classification of individuals according to certain shared characteristics.', 'Focus Groups; jury; teams', '{isa} Conceptual Entity; {inverse_isa} Professional or Occupational Group; {inverse_isa} Population Group; {inverse_isa} Family Group; {inverse_isa} Age Group; {inverse_isa} Patient or Disabled Group', 'A2.9', 'Few concepts will be assigned to this broad type.', 'grup', 1096, NULL, 'STY'),
(9, 'Living Beings', 'LIVE', 1111, 'Archaeon', 4, 'A member of one of the three domains of life, formerly called Archaebacteria under the taxon Bacteria, but now considered separate and distinct. Archaea are characterized by: 1) the presence of characteristic tRNAs and ribosomal RNAs; 2) the absence of peptidoglycan cell walls; 3) the presence of ether-linked lipids built from branched-chain subunits; and 4) their occurrence in unusual habitats. While archaea resemble bacteria in morphology and genomic organization, they resemble eukarya in their method of genomic replication.', 'Thermoproteales; Haloferax volcanii; Methanospirillum', '{isa} Organism', 'A1.1.1', NULL, 'arch', 1194, NULL, 'STY'),
(9, 'Living Beings', 'LIVE', 1112, 'Bacterium', 4, 'A small, typically one-celled, prokaryotic micro-organism.', 'Acetobacter; Bacillus cereus; Cytophaga', '{isa} Organism', 'A1.1.2', NULL, 'bact', 1007, NULL, 'STY'),
(9, 'Living Beings', 'LIVE', 1113, 'Eukaryote', 4, 'One of the three domains of life (the others being Bacteria and Archaea), also called Eukarya. These are organisms whose cells are enclosed in membranes and possess a nucleus. They comprise almost all multicellular and many unicellular organisms, and are traditionally divided into groups (sometimes called kingdoms) including Animals, Plants, Fungi, various Algae, and other taxa that were previously part of the old kingdom Protista.', 'Order Acarina; Bees; Plasmodium malariae', '{isa} Organism; {inverse_isa} Plant; {inverse_isa} Fungus; {inverse_isa} Animal', 'A1.1.3', NULL, 'euka', 1204, NULL, 'STY'),
(9, 'Living Beings', 'LIVE', 1114, 'Virus', 4, 'An organism consisting of a core of a single nucleic acid enclosed in a protective coat of protein. A virus may replicate only inside a host living cell. A virus exhibits some but not all of the usual characteristics of living things.', 'Coliphages; Echovirus; Parvoviridae', '{isa} Organism', 'A1.1.4', NULL, 'virs', 1005, NULL, 'STY'),
(9, 'Living Beings', 'LIVE', 1291, 'Professional or Occupational Group', 4, 'An individual or individuals classified according to their vocation.', 'Clergy; Demographers; Hospital Volunteers', '{isa} Group', 'A2.9.1', 'If the concept refers to the discipline or vocation itself, rather than to the individuals who have the vocation, then the type \'Occupation or Discipline\' will be assigned instead.', 'prog', 1097, NULL, 'STY'),
(9, 'Living Beings', 'LIVE', 1292, 'Population Group', 4, 'An indivdual or individuals classified according to their sex, racial origin, religion, common place of living, financial or social status, or some other cultural or behavioral attribute.', 'Asian Americans; Ethnic group; Adult Offenders', '{isa} Group', 'A2.9.2', NULL, 'popg', 1098, NULL, 'STY'),
(9, 'Living Beings', 'LIVE', 1293, 'Family Group', 4, 'An individual or individuals classified according to their family relationships or relative position in the family unit.', 'Daughter; Is an only child; Unmarried Fathers', '{isa} Group', 'A2.9.3', NULL, 'famg', 1099, NULL, 'STY'),
(9, 'Living Beings', 'LIVE', 1294, 'Age Group', 4, 'An individual or individuals classified according to their age.', 'Adult; Infant, Premature; Adolescent (age group)', '{isa} Group', 'A2.9.4', NULL, 'aggp', 1100, NULL, 'STY'),
(9, 'Living Beings', 'LIVE', 1295, 'Patient or Disabled Group', 4, 'An individual or individuals classified according to a disability, disease, condition or treatment.', 'Amputees; Institutionalized Child; Mentally Ill Persons', '{isa} Group', 'A2.9.5', NULL, 'podg', 1101, NULL, 'STY'),
(9, 'Living Beings', 'LIVE', 11131, 'Animal', 5, 'An organism with eukaryotic cells, and lacking stiff cell walls, plastids and photosynthetic pigments.', 'Animals; Animals, Laboratory; Carnivore', '{isa} Eukaryote; {inverse_isa} Vertebrate', 'A1.1.3.1', NULL, 'anim', 1008, NULL, 'STY'),
(9, 'Living Beings', 'LIVE', 11132, 'Fungus', 5, 'A eukaryotic organism characterized by the absence of chlorophyll and the presence of a rigid cell wall. Included here are both slime molds and true fungi such as yeasts, molds, mildews, and mushrooms.', 'Aspergillus clavatus; Blastomyces; Neurospora', '{isa} Eukaryote', 'A1.1.3.2', NULL, 'fngs', 1004, NULL, 'STY'),
(9, 'Living Beings', 'LIVE', 11133, 'Plant', 5, 'An organism having cellulose cell walls, growing by synthesis of inorganic substances, generally distinguished by the presence of chlorophyll, and lacking the power of locomotion. Plant parts are included here as well.', 'Aloe; Pollen; Helianthus species', '{isa} Eukaryote', 'A1.1.3.3', NULL, 'plnt', 1002, NULL, 'STY'),
(9, 'Living Beings', 'LIVE', 111311, 'Vertebrate', 6, 'An animal which has a spinal column.', 'Vertebrates; Gnathostomata vertebrate; Craniata <chordata>', '{isa} Animal; {inverse_isa} Amphibian; {inverse_isa} Bird; {inverse_isa} Fish; {inverse_isa} Reptile; {inverse_isa} Mammal', 'A1.1.3.1.1', 'Few concepts will be assigned to this broad type.', 'vtbt', 1010, NULL, 'STY'),
(9, 'Living Beings', 'LIVE', 1113111, 'Amphibian', 7, 'A cold-blooded, smooth-skinned vertebrate which characteristically hatches as an aquatic larva, breathing by gills. When mature, the amphibian breathes with lungs.', 'Salamandra; Urodela; Brazilian horned frog', '{isa} Vertebrate', 'A1.1.3.1.1.1', NULL, 'amph', 1011, NULL, 'STY'),
(9, 'Living Beings', 'LIVE', 1113112, 'Bird', 7, 'A vertebrate having a constant body temperature and characterized by the presence of feathers.', 'Serinus; Ducks; Quail', '{isa} Vertebrate', 'A1.1.3.1.1.2', NULL, 'bird', 1012, NULL, 'STY'),
(9, 'Living Beings', 'LIVE', 1113113, 'Fish', 7, 'A cold-blooded aquatic vertebrate characterized by fins and breathing by gills. Included here are fishes having either a bony skeleton, such as a perch, or a cartilaginous skeleton, such as a shark, or those lacking a jaw, such as a lamprey or hagfish.', 'Bass; Salmonidae; Whitefish', '{isa} Vertebrate', 'A1.1.3.1.1.3', NULL, 'fish', 1013, NULL, 'STY'),
(9, 'Living Beings', 'LIVE', 1113114, 'Mammal', 7, 'A vertebrate having a constant body temperature and characterized by the presence of hair, mammary glands and sweat glands.', 'Ursidae Family; Hamsters; Macaca', '{isa} Vertebrate; {inverse_isa} Human', 'A1.1.3.1.1.4', NULL, 'mamm', 1015, NULL, 'STY'),
(9, 'Living Beings', 'LIVE', 1113115, 'Reptile', 7, 'A cold-blooded vertebrate having an external covering of scales or horny plates. Reptiles breathe by means of lungs and are generally egg-laying.', 'Alligators; Water Mocassin; Genus Python (organism)', '{isa} Vertebrate', 'A1.1.3.1.1.5', NULL, 'rept', 1014, NULL, 'STY'),
(9, 'Living Beings', 'LIVE', 11131141, 'Human', 8, 'Modern man, the only remaining species of the Homo genus.', 'Homo sapiens; jean piaget; Member of public', '{isa} Mammal', 'A1.1.3.1.1.4.1', 'If a concept describes a human being from the point of view of occupational, family, social status, etc., then a type from the \'Group\' hierarchy will be assigned instead.', 'humn', 1016, NULL, 'STY'),
(10, 'Objects', 'OBJC', 1, 'Entity', 1, 'A broad type for grouping physical and conceptual entities.', 'Gifts, Financial; Image; Product Part', '{inverse_isa} Physical Object; {inverse_isa} Conceptual Entity', 'A', 'Few concepts will be assigned to this broad type.', 'enty', 1071, NULL, 'STY'),
(10, 'Objects', 'OBJC', 11, 'Physical Object', 2, 'An object perceptible to the sense of vision or touch.', 'Printed Media; Meteors; Physical object', '{isa} Entity; {inverse_isa} Organism; {inverse_isa} Anatomical Structure; {inverse_isa} Manufactured Object; {inverse_isa} Substance', 'A1', NULL, 'phob', 1072, NULL, 'STY'),
(10, 'Objects', 'OBJC', 113, 'Manufactured Object', 3, 'A physical object made by human beings.', 'car seat; Cooking and Eating Utensils; Goggles', '{isa} Physical Object; {inverse_isa} Medical Device; {inverse_isa} Research Device; {inverse_isa} Clinical Drug', 'A1.3', NULL, 'mnob', 1073, NULL, 'STY'),
(10, 'Objects', 'OBJC', 114, 'Substance', 3, 'A material with definite or fairly definite chemical composition.', 'Air (substance); Fossils; Plastics', '{isa} Physical Object; {inverse_isa} Body Substance; {inverse_isa} Chemical; {inverse_isa} Food', 'A1.4', NULL, 'sbst', 1167, NULL, 'STY'),
(10, 'Objects', 'OBJC', 1143, 'Food', 4, 'Any substance generally containing nutrients, such as carbohydrates, proteins, and fats, that can be ingested by a living organism and metabolized into energy and body tissue. Some foods are naturally occurring, others are either partially or entirely made by humans.', 'Beverages; Egg Yolk (Dietary); Ice Cream', '{isa} Substance', 'A1.4.3', 'Food additives, food preservatives, and food dyes should be given the type \'Chemical Viewed Functionally\'; \"Diet Coke\" would be assigned this type.', 'food', 1168, NULL, 'STY'),
(11, 'Occupations', 'OCCU', 126, 'Occupation or Discipline', 3, 'A vocation, academic discipline, or field of study, or a subpart of an occupation or discipline.', 'Aviation; Craniology; Ecology', '{isa} Conceptual Entity; {inverse_isa} Biomedical Occupation or Discipline', 'A2.6', 'If the concept refers to the individuals who have the vocation, the type \'Professional or Occupational Group\' will be assigned instead.', 'ocdi', 1090, NULL, 'STY'),
(11, 'Occupations', 'OCCU', 1261, 'Biomedical Occupation or Discipline', 4, 'A vocation, academic discipline, or field of study related to biomedicine.', 'Adolescent Medicine; Cellular Neurobiology; Dentistry', '{isa} Occupation or Discipline', 'A2.6.1', NULL, 'bmod', 1091, NULL, 'STY'),
(12, 'Organizations', 'ORGA', 127, 'Organization', 3, 'The result of uniting for a common purpose or function. The continued existence of an organization is not dependent on any of its members, its location, or particular facility. Components or subparts of organizations are also included here. Although the names of organizations are sometimes used to refer to the buildings in which they reside, they are not inherently physical in nature.', 'Labor Unions; United Nations; Boarding school', '{isa} Conceptual Entity; {inverse_isa} Health Care Related Organization; {inverse_isa} Professional Society; {inverse_isa} Self-help or Relief Organization', 'A2.7', NULL, 'orgt', 1092, NULL, 'STY'),
(12, 'Organizations', 'ORGA', 1271, 'Health Care Related Organization', 4, 'An established organization which carries out specific functions related to health care delivery or research in the life sciences.', 'Centers for Disease Control and Prevention (U.S.); Halfway Houses; Hospitals, Pediatric', '{isa} Organization', 'A2.7.1', 'Concepts for health care related professional societies are assigned the type \'Professional Society\'.', 'hcro', 1093, NULL, 'STY'),
(12, 'Organizations', 'ORGA', 1272, 'Professional Society', 4, 'An organization uniting those who have a common vocation or who are involved with a common field of study.', 'American Medical Association; International Council of Nurses; Library', '{isa} Organization', 'A2.7.2', NULL, 'pros', 1094, NULL, 'STY'),
(12, 'Organizations', 'ORGA', 1273, 'Self-help or Relief Organization', 4, 'An organization whose purpose and function is to provide assistance to the needy or to offer support to those sharing similar problems.', 'Alcoholics Anonymous; Charities - organization; Red Cross', '{isa} Organization', 'A2.7.3', NULL, 'shro', 1095, NULL, 'STY'),
(13, 'Phenomena', 'PHEN', 22, 'Phenomenon or Process', 2, 'A process or state which occurs naturally or as a result of an activity.', 'Disasters; Motor Traffic Accidents; Depolymerization', '{isa} Event; {inverse_isa} Injury or Poisoning; {inverse_isa} Human-caused Phenomenon or Process; {inverse_isa} Natural Phenomenon or Process', 'B2', NULL, 'phpr', 1067, NULL, 'STY'),
(13, 'Phenomena', 'PHEN', 221, 'Human-caused Phenomenon or Process', 3, 'A phenomenon or process that is a result of the activities of human beings.', 'Baby Boom; Cultural Evolution; Mass Media', '{isa} Phenomenon or Process; {inverse_isa} Environmental Effect of Humans', 'B2.1', 'If the concept refers to the activity itself, rather than the result of that activity, a type from the \'Activity\' hierarchy will be assigned instead.', 'hcpp', 1068, NULL, 'STY'),
(13, 'Phenomena', 'PHEN', 222, 'Natural Phenomenon or Process', 3, 'A phenomenon or process that occurs irrespective of the activities of human beings.', 'Air Movements; Corrosion; Lightning (phenomenon)', '{isa} Phenomenon or Process; {inverse_isa} Biologic Function', 'B2.2', NULL, 'npop', 1070, NULL, 'STY'),
(13, 'Phenomena', 'PHEN', 1221, 'Laboratory or Test Result', 4, 'The outcome of a specific test to measure an attribute or to determine the presence, absence, or degree of a condition.', 'Blood Flow Velocity; Serum Calcium Level; Spinal Fluid Pressure', '{isa} Finding', 'A2.2.1', 'Laboratory or test results are considered inherently quantitative and, thus, are not assigned the additional type \'Quantitative Concept\'.', 'lbtr', 1034, NULL, 'STY');
INSERT INTO `semantic_network` (`SemanticGroupCode`, `SemanticGroup`, `SemanticGroupAbr`, `CustomTreeNumber`, `SemanticTypeName`, `BranchPosition`, `Definition`, `Examples`, `RelationName`, `SemTypeTreeNo`, `UsageNote`, `Abbreviation`, `UniqueID`, `NonHumanFlag`, `RecordType`) VALUES
(13, 'Phenomena', 'PHEN', 2211, 'Environmental Effect of Humans', 4, 'A change in the natural environment that is a result of the activities of human beings.', 'Air Pollution; Desertification; Bioremediation', '{isa} Human-caused Phenomenon or Process', 'B2.1.1', NULL, 'eehu', 1069, NULL, 'STY'),
(13, 'Phenomena', 'PHEN', 2221, 'Biologic Function', 4, 'A state, activity or process of the body or one of its systems or parts.', 'Antibody Formation; Drug resistance; Homeostasis', '{isa} Natural Phenomenon or Process; {inverse_isa} Physiologic Function; {inverse_isa} Pathologic Function', 'B2.2.1', 'Few concepts will be assigned to this broad type.', 'biof', 1038, 'Y', 'STY'),
(14, 'Physiology', 'PHYS', 123, 'Organism Attribute', 3, 'A property of the organism or its major parts.', 'Age; Birth Weight; Eye Color', '{isa} Conceptual Entity; {inverse_isa} Clinical Attribute', 'A2.3', NULL, 'orga', 1032, 'Y', 'STY'),
(14, 'Physiology', 'PHYS', 1231, 'Clinical Attribute', 4, 'An observable or measurable property or state of an organism of clinical interest.', 'Bone Density; heart rate; Range of Motion, Articular', '{isa} Organism Attribute', 'A2.3.1', 'These are the attributes that are being evaluated or measured, not the results of the evaluation.', 'clna', 1201, NULL, 'STY'),
(14, 'Physiology', 'PHYS', 22211, 'Physiologic Function', 5, 'A normal process, activity, or state of the body.', 'Biorhythms; Hearing; Vasodilation', '{isa} Biologic Function; {inverse_isa} Organism Function; {inverse_isa} Organ or Tissue Function; {inverse_isa} Cell Function; {inverse_isa} Molecular Function', 'B2.2.1.1', NULL, 'phsf', 1039, NULL, 'STY'),
(14, 'Physiology', 'PHYS', 222111, 'Organism Function', 6, 'A physiologic function of the organism as a whole, of multiple organ systems, or of multiple organs or tissues.', 'Breeding; Hibernation; Motor Skills', '{isa} Physiologic Function; {inverse_isa} Mental Process', 'B2.2.1.1.1', NULL, 'orgf', 1040, NULL, 'STY'),
(14, 'Physiology', 'PHYS', 222112, 'Organ or Tissue Function', 6, 'A physiologic function of a particular organ, organ system, or tissue.', 'Osteogenesis; Renal Circulation; Tooth Calcification', '{isa} Physiologic Function', 'B2.2.1.1.2', NULL, 'ortf', 1042, NULL, 'STY'),
(14, 'Physiology', 'PHYS', 222113, 'Cell Function', 6, 'A physiologic function inherent to cells or cell components.', 'Cell Cycle; Cell division; Phagocytosis', '{isa} Physiologic Function', 'B2.2.1.1.3', NULL, 'celf', 1043, NULL, 'STY'),
(14, 'Physiology', 'PHYS', 222114, 'Molecular Function', 6, 'A physiologic function occurring at the molecular level.', 'Binding, Competitive; Electron Transport; Glycolysis', '{isa} Physiologic Function; {inverse_isa} Genetic Function', 'B2.2.1.1.4', NULL, 'moft', 1044, NULL, 'STY'),
(14, 'Physiology', 'PHYS', 2221111, 'Mental Process', 7, 'A physiologic function involving the mind or cognitive processing.', 'Anger; Auditory Fatigue; Avoidance Learning', '{isa} Organism Function', 'B2.2.1.1.1.1', NULL, 'menp', 1041, NULL, 'STY'),
(14, 'Physiology', 'PHYS', 2221141, 'Genetic Function', 7, 'Functions of or related to the maintenance, translation or expression of the genetic material.', 'Early Gene Transcription; Gene Amplification; RNA Splicing', '{isa} Molecular Function', 'B2.2.1.1.4.1', NULL, 'genf', 1045, NULL, 'STY'),
(15, 'Procedures', 'PROC', 2131, 'Health Care Activity', 4, 'An activity of or relating to the practice of medicine or involving the care of patients.', 'ambulatory care services; Clinic Activities; Preventive Health Services', '{isa} Occupational Activity; {inverse_isa} Laboratory Procedure; {inverse_isa} Diagnostic Procedure; {inverse_isa} Therapeutic or Preventive Procedure', 'B1.3.1', NULL, 'hlca', 1058, NULL, 'STY'),
(15, 'Procedures', 'PROC', 2132, 'Research Activity', 4, 'An activity carried out as part of research or experimentation.', 'Animal Experimentation; Biomedical Research; Experimental Replication', '{isa} Occupational Activity; {inverse_isa} Molecular Biology Research Technique', 'B1.3.2', 'In some cases, a concept may be assigned to both this type and the type \'Intellectual Product\'. For example, the concept \"Comparative Study\" might be viewed as both an activity and the result, or product, of that activity.', 'resa', 1062, NULL, 'STY'),
(15, 'Procedures', 'PROC', 2134, 'Educational Activity', 4, 'An activity related to the organization and provision of education.', 'Academic Training; Family Planning Training; Preceptorship', '{isa} Occupational Activity', 'B1.3.4', NULL, 'edac', 1065, NULL, 'STY'),
(15, 'Procedures', 'PROC', 21311, 'Laboratory Procedure', 5, 'A procedure, method, or technique used to determine the composition, quantity, or concentration of a specimen, and which is carried out in a clinical laboratory. Included here are procedures which measure the times and rates of reactions.', 'Blood Protein Electrophoresis; Crystallography; Radioimmunoassay', '{isa} Health Care Activity', 'B1.3.1.1', NULL, 'lbpr', 1059, NULL, 'STY'),
(15, 'Procedures', 'PROC', 21312, 'Diagnostic Procedure', 5, 'A procedure, method, or technique used to determine the nature or identity of a disease or disorder. This excludes procedures which are primarily carried out on specimens in a laboratory.', 'Biopsy; Heart Auscultation; Magnetic Resonance Imaging', '{isa} Health Care Activity', 'B1.3.1.2', NULL, 'diap', 1060, NULL, 'STY'),
(15, 'Procedures', 'PROC', 21313, 'Therapeutic or Preventive Procedure', 5, 'A procedure, method, or technique designed to prevent a disease or a disorder, or to improve physical function, or used in the process of treating a disease or injury.', 'Cesarean section; Dermabrasion; Family psychotherapy', '{isa} Health Care Activity', 'B1.3.1.3', NULL, 'topp', 1061, NULL, 'STY'),
(15, 'Procedures', 'PROC', 21321, 'Molecular Biology Research Technique', 5, 'Any of the techniques used in the study of or the directed modification of the gene complement of a living organism.', 'Northern Blotting; Genetic Engineering; In Situ Hybridization', '{isa} Research Activity', 'B1.3.2.1', NULL, 'mbrt', 1063, NULL, 'STY');
'''
